+++ uikit/src/main/java/com/sendbird/uikit/activities/viewholder/MessageViewHolderFactory.java
@@ -34,6 +34,7 @@ import com.sendbird.uikit.databinding.SbViewOtherVoiceMessageBinding;
 import com.sendbird.uikit.databinding.SbViewParentMessageInfoHolderBinding;
 import com.sendbird.uikit.databinding.SbViewSuggestedRepliesMessageBinding;
 import com.sendbird.uikit.databinding.SbViewTimeLineMessageBinding;
+import com.sendbird.uikit.databinding.SbViewTypingIndicatorMessageBinding;
 import com.sendbird.uikit.internal.extensions.MessageExtensionsKt;
 import com.sendbird.uikit.internal.ui.viewholders.AdminMessageViewHolder;
 import com.sendbird.uikit.internal.ui.viewholders.FormMessageViewHolder;
@@ -57,9 +58,11 @@ import com.sendbird.uikit.internal.ui.viewholders.OtherVoiceMessageViewHolder;
 import com.sendbird.uikit.internal.ui.viewholders.ParentMessageInfoViewHolder;
 import com.sendbird.uikit.internal.ui.viewholders.SuggestedRepliesViewHolder;
 import com.sendbird.uikit.internal.ui.viewholders.TimelineViewHolder;
+import com.sendbird.uikit.internal.ui.viewholders.TypingIndicatorViewHolder;
 import com.sendbird.uikit.model.MessageListUIParams;
 import com.sendbird.uikit.model.SuggestedRepliesMessage;
 import com.sendbird.uikit.model.TimelineMessage;
+import com.sendbird.uikit.model.TypingIndicatorMessage;
 import com.sendbird.uikit.utils.MessageUtils;
 
 import java.util.Map;
@@ -228,6 +231,9 @@ public class MessageViewHolderFactory {
             case VIEW_TYPE_FORM_TYPE_MESSAGE:
                 holder = new FormMessageViewHolder(SbViewFormMessageBinding.inflate(inflater, parent, false), messageListUIParams);
                 break;
+            case VIEW_TYPE_TYPING_INDICATOR:
+                holder = new TypingIndicatorViewHolder(SbViewTypingIndicatorMessageBinding.inflate(inflater, parent, false), messageListUIParams);
+                break;
             default:
                 // unknown message type
                 if (viewType == MessageType.VIEW_TYPE_UNKNOWN_MESSAGE_ME) {
@@ -323,6 +329,8 @@ public class MessageViewHolderFactory {
             type = MessageType.VIEW_TYPE_ADMIN_MESSAGE;
         } else if (message instanceof SuggestedRepliesMessage) {
             type = MessageType.VIEW_TYPE_SUGGESTED_REPLIES;
+        }  else if (message instanceof TypingIndicatorMessage) {
+            type = MessageType.VIEW_TYPE_TYPING_INDICATOR;
         } else {
             if (MessageUtils.isMine(message)) {
                 type = MessageType.VIEW_TYPE_UNKNOWN_MESSAGE_ME;